# lib/api-zod/src/generated/types/copilotMessage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * StadiumSense AI — FIFA World Cup 2026 Stadium Operations API
 * OpenAPI spec version: 0.1.0
 */
import type { CopilotMessageRole } from './copilotMessageRole';

export interface CopilotMessage {
  role: CopilotMessageRole;
  content: string;
}