# lib/api-zod/src/generated/types/crowdAnalysisInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * StadiumSense AI — FIFA World Cup 2026 Stadium Operations API
 * OpenAPI spec version: 0.1.0
 */
import type { CrowdAnalysisInputMatchPhase } from './crowdAnalysisInputMatchPhase';
import type { CrowdGateSnapshot } from './crowdGateSnapshot';

export interface CrowdAnalysisInput {
  gates: CrowdGateSnapshot[];
  matchPhase: CrowdAnalysisInputMatchPhase;
}